Обновление века часть 1 доработки
