+++ src/presenter/film-list-presenter.js
@@ -1,119 +1,185 @@
-import { KeysName } from '../render.js';
-import FilmCardView from '../view/film-card-view';
-import FilmPopupView from '../view/film-popup-view';
-import {FilmActionType} from '../render.js';
+import {FilmListNames} from '../render.js';
+import FilmListView from '../view/film-list-view.js';
+import { FilmPresenter } from './film-card-presenter.js';
+import FilmsView from '../view/films-view.js';
+import NoFilmsView from '../view/no-films-view.js';
 import {render,
   RenderPosition,
   remove,
   replace} from '../render.js';
+import { SortTypes } from '../render.js';
+import ShowButtonView from '../view/show-button-view.js';
+import SortView from '../view/sort-view.js';
+import { sortRatingDown,
+  sortCommentCountDown,
+  sortReleaseDateDown } from '../utils/film-sort.js';
+import { updateItem } from '../utils/common.js';
 
-export class FilmPresenter {
-#container = null;
-#changeData = null;
-#film  = null;
-#filmCardComponent = null;
-#popupComponent = null;
-#siteFooter = document.querySelector('.footer');
-_callback = {};
+const FILM_COUNT_PER_STEP = 5;
+const FILM_TOP_RATED_COUNT = 2;
+const FILM_MOST_COMMENTED_COUNT = 2;
 
-constructor(container, changeData) {
-  this.#container = container;
-  this.#changeData = changeData;
-}
+export default class FilmListPresenter {
+  #container = null;
 
-init = (film) => {
-  this.#film = film;
+  #filmListComponent = new FilmListView();
+  #filmsComponent = new FilmsView();
+  #sortComponent = new SortView();
+  #showMoreComponent = new ShowButtonView();
+  #filmsList = [];
 
-  const prevFilmComponent = this.#filmCardComponent;
-  this.#filmCardComponent = new FilmCardView(film);
+  #sourceFilmList = [];
+  #renderedFilmCount = FILM_COUNT_PER_STEP;
+  #filmPresenter = new Map();
+  #filmPresenerWithPopup = null;
+  #currentSortType = SortTypes.DEFAULT;
 
-  this.#filmCardComponent.setClickHandler(this.#handleCardClick);
-  this.#filmCardComponent.setActionHandler(this.#handlerFilmAction);
-
-  if(prevFilmComponent === null) {
-    render(this.#container, this.#filmCardComponent);
-    return;
+  constructor(container) {
+    this.#container = container;
   }
 
-  replace(this.#filmCardComponent, prevFilmComponent);
-  remove(prevFilmComponent);
-}
+  init = (filmList) => {
+    this.#filmsList = [...filmList];
+    this.#sourceFilmList = [...filmList];
 
-openPopup = () => {
-  const prevPopupComponent = this.#popupComponent;
-  this.#popupComponent = new FilmPopupView(this.#film);
-
-  this.#popupComponent.setCloseClickHandler(this.#handleClosePopup);
-  this.#popupComponent.setActionHandler(this.#handlerFilmAction);
-
-  document.addEventListener('keydown', this.#onEscKeyDown);
-  document.body.classList.add('hide-overflow');
-
-  render(this.#siteFooter, this.#popupComponent, RenderPosition.AFTEREND);
-
-  if(prevPopupComponent === null) {
-    render(this.#siteFooter, this.#popupComponent, RenderPosition.AFTEREND);
-    return;
+    this.#renderSort();
+    this.#renderSectionFilms();
   }
 
-  replace(this.#popupComponent, prevPopupComponent);
-  remove(prevPopupComponent);
-}
+  #sortFilmList = (sortType) => {
+    switch (sortType) {
+      case SortTypes.DATE:
+        this.#filmsList.sort(sortReleaseDateDown);
+        break;
+      case SortTypes.RATING:
+        this.#filmsList.sort(sortRatingDown);
+        break;
+      default:
+        this.#filmsList = [...this.#sourceFilmList];
+    }
 
-removePopup() {
-  document.removeEventListener('keydown', this.#onEscKeyDown);
-  document.body.classList.remove('hide-overflow');
-
-  if(this.#popupComponent !== null) {
-    remove(this.#popupComponent);
-    this.#popupComponent = null;
+    this.#currentSortType = sortType;
   }
-}
 
-#onEscKeyDown = (evt) => {
-  if (evt.key === KeysName.ESC || evt.key === KeysName.ESCAPE) {
-    evt.preventDefault();
-    this.removePopup();
+  #handleSortTypeChange = (sortType) => {
+    if (this.#currentSortType === sortType) {
+      return;
+    }
+
+    this.#reRenderSort(sortType);
+
+    this.#sortFilmList(sortType);
+    this.#clearFilmList();
+    this.#initFilmList();
   }
-}
 
-setCardClick = (callback) => {
-  this._callback.cardClick =  callback;
-}
+  #reRenderSort = (sortType) =>  {
+    const updatedSortComponent = new SortView(sortType);
+    replace(updatedSortComponent, this.#sortComponent);
+    updatedSortComponent.setSortTypeChangeHandler(this.#handleSortTypeChange);
+    this.#sortComponent = updatedSortComponent;
+  }
 
-setCardClose = (callback) => {
-  this._callback.cardClose =  callback;
-}
+  #renderSort = () => {
+    render(this.#container, this.#sortComponent);
+    this.#sortComponent.setSortTypeChangeHandler(this.#handleSortTypeChange);
+  }
 
-#handleCardClick = () => {
-  this._callback.cardClick();
-}
+  #handleFilmChange = (updatedFilm) => {
+    this.#filmsList = updateItem(this.#filmsList, updatedFilm);
+    this.#sourceFilmList = updateItem(this.#sourceFilmList, updatedFilm);
 
-#handleClosePopup = () => {
-  this.removePopup();
-  this._callback.cardClose();
-}
+    this.#filmPresenter.get(FilmListNames.ALL_FILMS + updatedFilm.id)?.init(updatedFilm);
+    this.#filmPresenter.get(FilmListNames.TOP_RATED + updatedFilm.id)?.init(updatedFilm);
+    this.#filmPresenter.get(FilmListNames.MOST_COMMENTED + updatedFilm.id)?.init(updatedFilm);
 
-#handlerFilmAction = (type) => {
-  switch (type) {
-    case FilmActionType.ADD_WATCH_LIST:
-      this.#changeData({...this.#film, isWatchList: !this.#film.isWatchList});
-      break;
-
-    case FilmActionType.MARK_WATCHED:
-      this.#changeData({...this.#film, isWatched: !this.#film.isWatched});
-      break;
-
-    case FilmActionType.MARK_FAVORITE:
-      this.#changeData({...this.#film, isFavorite: !this.#film.isFavorite});
-      break;
+    if(this.#filmPresenerWithPopup) {
+      this.#filmPresenerWithPopup.openPopup();
+    }
 
   }
 
-}
+  #handleCardClick = (filmPresenter) => {
+    if(this.#filmPresenerWithPopup) {
+      this.#filmPresenerWithPopup.removePopup();
+    }
+    filmPresenter.openPopup();
+    this.#filmPresenerWithPopup = filmPresenter;
+  };
 
-destroy = () => {
-  remove(this.#filmCardComponent);
-}
+  #handleCardClose = () => {
+    this.#filmPresenerWithPopup = null;
+  }
 
+  #renderSectionFilms = () => {
+    render(this.#container, this.#filmListComponent);
+    this.#initFilmList();
+  }
+
+  #renderFilmCard = (containerId, film) => {
+    const container = this.#filmListComponent.getFilmList(containerId);
+    const filmPresenter = new FilmPresenter(container, this.#handleFilmChange);
+    filmPresenter.setCardClick(() => this.#handleCardClick(filmPresenter));
+    filmPresenter.setCardClose(this.#handleCardClose);
+    filmPresenter.init(film);
+
+    this.#filmPresenter.set(containerId + film.id, filmPresenter);
+  };
+
+  #renderFilms = (containerId, films, from, to) => {
+    if(films.length === 0) {
+      render(this.#container, new NoFilmsView());
+    } else {
+      films
+        .slice(from, to)
+        .forEach((film) => this.#renderFilmCard(containerId, film));
+    }
+  }
+
+  #renderMainFilmList = (films) => {
+    const from = 0;
+    const to = Math.min(films.length, FILM_COUNT_PER_STEP);
+
+    this.#renderFilms(FilmListNames.ALL_FILMS, films, from, to);
+
+    if (this.#filmsList.length > FILM_COUNT_PER_STEP) {
+      this.#renderMoreButton();
+    }
+  }
+
+  get filmsOrderedByTopRated() {
+    return [...this.#filmsList].sort(sortRatingDown);
+  }
+
+  get filmsOrderedByCommentCount() {
+    return [...this.#filmsList].sort(sortCommentCountDown);
+  }
+
+  #handleMoreButtonClick = () => {
+    const from = this.#renderedFilmCount;
+    const to = this.#renderedFilmCount + FILM_COUNT_PER_STEP;
+    this.#renderFilms(FilmListNames.ALL_FILMS, this.#filmsList, from, to);
+    this.#renderedFilmCount = to;
+
+    if (this.#renderedFilmCount >= this.#filmsList.length) {
+      remove(this.#showMoreComponent);
+    }
+  }
+
+  #renderMoreButton = () => {
+    render(this.#filmListComponent.allListElement, this.#showMoreComponent, RenderPosition.AFTEREND);
+    this.#showMoreComponent.setClickHandler(this.#handleMoreButtonClick);
+  }
+
+  #initFilmList = () => {
+    this.#renderMainFilmList(this.#filmsList);
+    this.#renderFilms(FilmListNames.TOP_RATED, this.filmsOrderedByTopRated, 0, FILM_TOP_RATED_COUNT);
+    this.#renderFilms(FilmListNames.MOST_COMMENTED, this.filmsOrderedByCommentCount, 0, FILM_MOST_COMMENTED_COUNT);
+  };
+
+  #clearFilmList = () => {
+    this.#filmPresenter.forEach((presenter) => presenter.destroy());
+    this.#filmPresenter.clear();
+    this.#renderedFilmCount = FILM_COUNT_PER_STEP;
+  }
 }
